refactor: move pipeline.py to top-level, fix docstring (#319)
Pipeline is orchestration logic that consumes Shim config — it belongs
as a peer to shims/ and converters/, not inside the shims/ package.

- Move shims/pipeline.py → pipeline.py (top-level module)
- Old path (shims/pipeline.py) becomes backward-compat re-export
- Fix apply_shim_to_ir docstring: "always use the return value"
- Update gateway/proxy.py and tests to import from new location
